Plumb HAVE_DGL to compiler + widen barbox coords to Point<double>
Two unrelated fixes for the DPF bump.

1. Pass -DHAVE_DGL to CXX in DGL_FLAGS.

   Current DPF's DistrhoPluginChecks.h silently forces
   DISTRHO_PLUGIN_HAS_UI to 0 when the compiler doesn't see
   -DHAVE_DGL, which then makes DistrhoUIMain.cpp emit
       # error Trying to build UI without DISTRHO_PLUGIN_HAS_UI
                set to 1
   Our fork has HAVE_DGL as a make variable used in ifeq
   conditionals but never forwarded the define to the compiler;
   old DPF didn't require it. Add it next to -DDGL_OPENGL /
   -DDGL_CAIRO, matching upstream Makefile.plugins.mk:228,240.

2. Widen barbox.hpp coordinates to Point<double>.

   DPF's event 'pos' is now Point<double> and Point lacks a
   cross-type conversion constructor, so the 'anchor = ev.pos'
   and 'mousePosition = ev.pos' assignments stopped compiling.
   Cascade the change to all related signatures (calcIndex,
   setStateFromPosition, setStateFromLine, setValueFromPosition,
   setValueFromLine) and the one explicit Point<int>(...)
   construction. The bodies already cast to int where they need
   discrete indexing, so the widening is harmless.

+++ plugin/dpf/common/gui/barbox.hpp
@@ -46,8 +46,8 @@ private:
 
   Scale &scale;
 
-  Point<int> mousePosition{-1, -1};
-  Point<int> anchor{0, 0};
+  Point<double> mousePosition{-1, -1};
+  Point<double> anchor{0, 0};
   BarState anchorState = BarState::active;
   int indexL = 0;
   int indexR = 0;
@@ -632,7 +632,7 @@ public:
         setStateFromLine(anchor, ev.pos, anchorState);
       } else if (ev.mod & kModifierShift) {
         mousePosition
-          = Point<int>(anchor.getX(), std::clamp<int>(ev.pos.getY(), 0, getHeight()));
+          = Point<double>(anchor.getX(), std::clamp<double>(ev.pos.getY(), 0.0, double(getHeight())));
         setValueFromPosition(mousePosition, 0);
       } else {
         setValueFromLine(anchor, ev.pos, ev.mod);
@@ -676,7 +676,7 @@ public:
   }
 
 private:
-  inline size_t calcIndex(Point<int> position)
+  inline size_t calcIndex(Point<double> position)
   {
     return size_t(indexL + position.getX() / sliderWidth);
   }
@@ -699,7 +699,7 @@ private:
     return idx < snapValue.size() ? snapValue[idx] : 1.0;
   }
 
-  BarState setStateFromPosition(Point<int> position, BarState state)
+  BarState setStateFromPosition(Point<double> position, BarState state)
   {
     size_t index = calcIndex(position);
     if (index >= value.size()) return BarState::active;
@@ -708,7 +708,7 @@ private:
     return barState[index];
   }
 
-  void setStateFromLine(Point<int> p0, Point<int> p1, BarState state)
+  void setStateFromLine(Point<double> p0, Point<double> p1, BarState state)
   {
     if (p0.getX() > p1.getX()) std::swap(p0, p1);
 
@@ -728,7 +728,7 @@ private:
     repaint();
   }
 
-  void setValueFromPosition(Point<int> position, uint modifier)
+  void setValueFromPosition(Point<double> position, uint modifier)
   {
     size_t index = calcIndex(position);
     if (index >= value.size()) return;
@@ -748,7 +748,7 @@ private:
     repaint();
   }
 
-  void setValueFromLine(Point<int> p0, Point<int> p1, uint modifier)
+  void setValueFromLine(Point<double> p0, Point<double> p1, uint modifier)
   {
     if (p0.getX() > p1.getX()) std::swap(p0, p1);
